% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reps.R
\name{add_reps}
\alias{add_reps}
\title{Add replicates}
\usage{
add_reps(df, sep = "_")
}
\arguments{
\item{df}{A dataframe containing two columns for well IDs and Sample IDs}

\item{sep}{a character string to separate the terms.}
}
\value{
A dataframe with replicate numbers pasted to the Sample IDs
}
\description{
Adds replicate information to the sample IDs. Well IDs should be formatted
like so: A4, B9, H11, J24
}
