% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MS.R
\name{calculate_MS}
\alias{calculate_MS}
\title{Calculate Maximum Slope}
\usage{
calculate_MS(
  data,
  col = lifecycle::deprecated(),
  .by = lifecycle::deprecated(),
  by = "well"
)
}
\arguments{
\item{data}{A dataframe containing real-time reads. It is recommended to use a dataframe made from normalize_RFU.}

\item{col}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{.by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use "by" instead.}

\item{by}{Grouping factor. Should typically be by individual wells.}
}
\value{
A dataframe containing the real-time slope values as change in RFU/sec.
}
\description{
Uses a sliding window to calculate the slope of real-time reads.
}
