% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_threshold.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Calculate a Threshold for Rate Determination}
\usage{
calculate_threshold(
  data,
  values = "rfu",
  time = lifecycle::deprecated(),
  background_time = 0,
  method = c("stdev"),
  multiplier = 1
)
}
\arguments{
\item{data}{A dataframe output from get_real.}

\item{values}{Column containing fluorescent values.}

\item{time}{Column containing time values.}

\item{background_time}{Float; the time point used for background fluorescence.}

\item{method}{Method for determining threshold; default is "stdev".}

\item{multiplier}{For some methods, will add a multiplier for more conservative thresholds.}
}
\value{
A float value.
}
\description{
Calculates a threshold for determining time-to-threshold and rate of amyloid formation.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
threshold <- get_quic(file) |>
  calculate_threshold(multiplier=10)

}
