% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tables.R
\name{convert_tables}
\alias{convert_tables}
\title{Convert tables into a single column in a dataframe.}
\usage{
convert_tables(tab, na_omit = TRUE)
}
\arguments{
\item{tab}{A table/matrix or a list of tables/matrices.}

\item{na_omit}{Logical; if true, will remove rows with NA.}
}
\value{
A dataframe column.
}
\description{
Accepts a table or matrix or a list of tables and matrices
and converts them into dataframe columns.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
tabs <- organize_tables(file)
convert_tables(tabs)

}
