% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Retrieve the BMG metadata}
\usage{
get_meta(file)
}
\arguments{
\item{file}{The Excel file exported from MARS.}
}
\value{
A dataframe containing the Meta_ID and Meta_info
}
\description{
Takes the Excel file exported from MARS and compiles the metadata in the
header.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_meta(file)

}
