% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quic.R
\name{get_quic}
\alias{get_quic}
\title{Get all time-series and metadata.}
\usage{
get_quic(
  file,
  transpose_table = lifecycle::deprecated(),
  norm_point = 2,
  which_table = 1,
  window_size = 2,
  smooth = FALSE,
  smooth_factor = 4,
  zero = FALSE,
  .by = lifecycle::deprecated(),
  by = c("well"),
  plate = 96,
  sheet = 2
)
}
\arguments{
\item{file}{An Excel file exported by BMG.}

\item{transpose_table}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument no longer has any effect.}

\item{norm_point}{Integer, defines the cycle to use as background fluorescence.}

\item{which_table}{Integer, defines which table in the Excel sheet contains the real-time data. Should usually be set to 1.}

\item{window_size}{Integer, defines the window size for estimating the derivative.}

\item{smooth}{Logical; if true, will smooth the data using a rolling mean.}

\item{smooth_factor}{Integer; the window size for smoothing.}

\item{zero}{Logical; if true, will zero out the background fluorescence.}

\item{.by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use "by" instead.}

\item{by}{Grouping factor. Should typically be by 'well'. Can also include 'sample' and 'dilution' if present.}

\item{plate}{Integer; either 96 or 384 to denote the type of well plate being used.}

\item{sheet}{Integer; the sheet number to read from.}
}
\value{
An S3 object of class "quic" containing all time-series data and sample/plate metadata.
}
\description{
Accepts an Excel file or a data frame of real-time RT-QuIC data.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_quic(file)

}
