% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real.R
\name{get_real}
\alias{get_real}
\title{Get Real-Time RT-QuIC Fluorescence Data}
\usage{
get_real(
  data,
  order_by_sample = lifecycle::deprecated(),
  transpose_table = lifecycle::deprecated()
)
}
\arguments{
\item{data}{Either an Excel file or a dataframe.}

\item{order_by_sample}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument no longer has any effect.}

\item{transpose_table}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument no longer has any effect.}
}
\value{
A list of dataframes containing the formatted real-time data.
}
\description{
Accepts an Excel file or a dataframe of real-time RT-QuIC data.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_real(file)

}
