% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_tables.R
\name{organize_tables}
\alias{organize_tables}
\title{Organize MARS Tables}
\usage{
organize_tables(file, plate = 96)
}
\arguments{
\item{file}{An Excel file exported from MARS.}

\item{plate}{Integer either 96 or 384 to denote microplate type.}
}
\value{
A list containing tibbles.
}
\description{
Extracts the tables from the microplate view sheet in the MARS Excel file
and adds each table to a list.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
organize_tables(file)

}
