% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_view.R
\name{plate_view}
\alias{plate_view}
\alias{plate_view.quic}
\alias{plate_view.data.frame}
\title{Real-Time Plate View}
\usage{
plate_view(data, ...)

\method{plate_view}{quic}(data, plate = data$params$plate, ...)

\method{plate_view}{data.frame}(
  data,
  plate = 96,
  sep = "\\n",
  plot_deriv = TRUE,
  flu_color = "black",
  der_color = "blue",
  ...
)
}
\arguments{
\item{data}{Either an S3 object of class "quic" returned by \code{\link[=get_quic]{get_quic()}}, or a
long-format data frame with \code{well}, \code{sample}, \code{dilution}, \code{time}, \code{norm}
and \code{deriv} columns.}

\item{...}{Arguments passed on to the data frame method.}

\item{plate}{Integer either 96 or 384 to denote microplate type. For "quic"
objects, defaults to the plate size recorded by \code{\link[=get_quic]{get_quic()}}.}

\item{sep}{A string defining how sample IDs and dilutions should be separated.}

\item{plot_deriv}{Logical; should the derivative be plotted?}

\item{flu_color}{Line color of the fluorescent data.}

\item{der_color}{Line color of the derivative data.}
}
\value{
A ggplot object
}
\description{
Converts the real-time data into a ggplot figure. The layout is either 8x12
or 16x24 for 96- and 384-well plates, respectively.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)

get_quic(file) |>
  plate_view(plot_deriv = FALSE)

}
