% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\alias{plot_metrics.quic_metrics}
\alias{plot_metrics.data.frame}
\title{Plot metrics generated from the "calculate" family of quicR functions.}
\usage{
plot_metrics(data, ...)

\method{plot_metrics}{quic_metrics}(data, ...)

\method{plot_metrics}{data.frame}(
  data,
  ...,
  sample_col = "sample",
  color = "black",
  fill = "dilution",
  dilution_bool = TRUE,
  nrow = 2,
  ncol = 2
)
}
\arguments{
\item{data}{Either an S3 object of class "quic_metrics" returned by
\code{\link[=calculate_metrics]{calculate_metrics()}}, or a data frame of metrics.}

\item{...}{A list of faceting columns. If empty, groups are set to "mpr", "ms", "ttt", and "raf".}

\item{sample_col}{The name of the column containing the sample IDs.}

\item{color}{The color used for the boxplot outline.}

\item{fill}{The column containing the fill aesthetic. Usually the dilution column.}

\item{dilution_bool}{Logical; should dilution factors be included in the plot?}

\item{nrow}{Integer; number of rows to output in the plot.}

\item{ncol}{Integer; number of columns to output in the plot.}
}
\value{
A ggplot object
}
\description{
Generates a faceted figure of boxplots.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)

get_quic(file) |>
  calculate_metrics(threshold = 3) |>
  plot_metrics()

}
