% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quic-methods.R
\name{quic-coercion}
\alias{quic-coercion}
\alias{as_tibble.quic}
\alias{as.data.frame.quic}
\alias{as_tibble.quic_metrics}
\alias{as.data.frame.quic_metrics}
\title{Convert quicR objects to data frames}
\usage{
\method{as_tibble}{quic}(x, ...)

\method{as.data.frame}{quic}(x, row.names = NULL, optional = FALSE, ...)

\method{as_tibble}{quic_metrics}(x, ...)

\method{as.data.frame}{quic_metrics}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class "quic" or "quic_metrics".}

\item{...}{Ignored.}

\item{row.names, optional}{Ignored; included for compatibility with the generic.}
}
\value{
A data frame (or tibble, for \code{as_tibble()}).
}
\description{
Flattens a "quic" object into the long-format data frame (one row per well
per read), or extracts the metrics table from a "quic_metrics" object.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
x <- get_quic(file)
head(as.data.frame(x))

}
