% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quic-methods.R
\name{quic-plot}
\alias{quic-plot}
\alias{autoplot.quic}
\alias{autoplot.quic_metrics}
\alias{plot.quic}
\alias{plot.quic_metrics}
\title{Plot quicR objects}
\usage{
\method{autoplot}{quic}(object, ...)

\method{autoplot}{quic_metrics}(object, ...)

\method{plot}{quic}(x, y, ...)

\method{plot}{quic_metrics}(x, y, ...)
}
\arguments{
\item{object, x}{An object of class "quic" or "quic_metrics".}

\item{...}{Passed on to \code{\link[=plate_view]{plate_view()}} or \code{\link[=plot_metrics]{plot_metrics()}}.}

\item{y}{Ignored; included for compatibility with the generic.}
}
\value{
A ggplot object (invisibly, for \code{plot()}).
}
\description{
\code{autoplot()} returns a ggplot: a \code{\link[=plate_view]{plate_view()}} for "quic" objects and a
\code{\link[=plot_metrics]{plot_metrics()}} boxplot for "quic_metrics" objects. \code{plot()} draws that
same plot and returns it invisibly.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
x <- get_quic(file)
plot(x, plot_deriv = FALSE)
plot(calculate_metrics(x, threshold = 3))

}
