% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quic-methods.R
\name{quic-print}
\alias{quic-print}
\alias{print.quic}
\alias{print.quic_metrics}
\alias{print.summary.quic}
\alias{print.summary.quic_metrics}
\title{Print quicR objects}
\usage{
\method{print}{quic}(x, n = 6, ...)

\method{print}{quic_metrics}(x, n = 6, ...)

\method{print}{summary.quic}(x, ...)

\method{print}{summary.quic_metrics}(x, ...)
}
\arguments{
\item{x}{An object of class "quic", "quic_metrics", or one of their summaries.}

\item{n}{Integer; the number of rows of data to show.}

\item{...}{Passed on to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print quicR objects
}
