% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicR.R
\docType{package}
\name{quicR-package}
\alias{quicR}
\alias{quicR-package}
\title{@description
Data handling for real-time quaking induced conversion assays}
\description{
See the README on \href{https://github.com/gage1145/quicR}{Github}
}
\details{
@keywords internal
}
\author{
\strong{Maintainer}: Gage Rowden \email{gage.rowden1145@gmail.com} (\href{https://orcid.org/0000-0002-7517-0480}{ORCID})

}
