% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_raw.R
\name{separate_raw}
\alias{separate_raw}
\title{Separate Real-Time Data into separate dataframes.}
\usage{
separate_raw(file, num_rows, export_name)
}
\arguments{
\item{file}{An Excel file exported from MARS.}

\item{num_rows}{Number of rows in the header to ignore.}

\item{export_name}{The name of the original file or an orignal name.}
}
\value{
An Excel file with separated raw real-time data.
}
\description{
If multiple real-time reads were exported from MARS, separate_raw will parse
them out and separate them. It will also export to an Excel file with each
real-time data having its own sheet.
}
