% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_real.R
\name{transpose_real}
\alias{transpose_real}
\title{Transpose Real-Time Data}
\usage{
transpose_real(data)
}
\arguments{
\item{data}{A dataframe generated from get_real.}
}
\value{
A transposed dataframe containing real-time normalized fluorescence values.
}
\description{
Transposes the real-time data table exported by the BMG software. Accepts
output from the function, "get_real".
}
