% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMG_format.R
\name{BMG_format}
\alias{BMG_format}
\title{Format Table for BMG Sample ID Import}
\usage{
BMG_format(file)
}
\arguments{
\item{file}{A .CSV file containing the plate layout of Sample IDs.}
}
\value{
A text file containing information for import into the BMG control software.
}
\description{
BMG_format accepts a plate layout .CSV file and formats the Sample IDs into
a format which can be easily imported into the BMG control software.
}
