% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC.R
\name{calculate_AUC}
\alias{calculate_AUC}
\title{Calculate the Area Under the Curve}
\usage{
calculate_AUC(data, x = "Time", y = "Norm", .by = "Wells")
}
\arguments{
\item{data}{A data frame output from 'get_quic()'.}

\item{x}{The column name containing the time data.}

\item{y}{The column containing the normalized fluorescence data.}

\item{.by}{Grouping factor. Should typically be by individual wells.}
}
\value{
A data frame containing well-matched AUC values.
}
\description{
Maxpoint ratio is defined as the maximum relative fluorescence divided by the
background fluorescence.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_quic(file) |>
  calculate_AUC()

}
