% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Generate a data frame with calculated metrics.}
\usage{
calculate_metrics(
  data,
  ...,
  threshold = 2,
  time_col = lifecycle::deprecated(),
  ttt_values = lifecycle::deprecated(),
  auc_values = lifecycle::deprecated(),
  norm_col = lifecycle::deprecated(),
  deriv_col = lifecycle::deprecated(),
  flip_ratio = FALSE
)
}
\arguments{
\item{data}{A S3 object of class "quic".}

\item{...}{A list of grouping factors. If left empty, function groups by "Sample IDs", "Dilutions", and "Wells".}

\item{threshold}{Float; the threshold applied to the calculation of time-to-threshold.}

\item{time_col}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{ttt_values}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{auc_values}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{norm_col}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{deriv_col}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{flip_ratio}{Logical; should the quenching ratio be calculated as max / last (default) or last / max?}
}
\value{
An S3 object of class "quic_metrics" containing relevant kinetic metrics.
}
\description{
Uses functions from the "calculate" family of quicR functions to generate an analyzed dataframe.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
get_quic(file) |>
 calculate_metrics(threshold = 3)

}
