% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Generate a data frame with calculated metrics.}
\usage{
calculate_metrics(
  data,
  ...,
  threshold = 2,
  time_col = "Time",
  ttt_values = "Norm",
  auc_values = "Norm",
  norm_col = "Norm",
  deriv_col = "Deriv",
  flip_ratio = FALSE
)
}
\arguments{
\item{data}{A data frame containing the raw RT-QuIC data.}

\item{...}{A list of grouping factors. If left empty, function groups by "Sample IDs", "Dilutions", and "Wells".}

\item{threshold}{Float; the threshold applied to the calculation of time-to-threshold.}

\item{time_col}{String; column name containing the time values.}

\item{ttt_values}{String; column name containing values to use for calculating time-to-threshold.}

\item{auc_values}{String; column name containing values to use for calculating the area under the curve.}

\item{norm_col}{String; column name containing the normalized fluorescent values.}

\item{deriv_col}{String; column name containing the estimated derivative values.}

\item{flip_ratio}{Logical; should the quenching ratio be calculated as max / last (default) or last / max?}
}
\value{
A data frame of calculated metrics.
}
\description{
Uses functions from the "calculate" family of quicR functions to generate an analyzed dataframe.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
get_quic(file) |>
 calculate_metrics()

}
