% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Generate a dataframe with calculated metrics.}
\usage{
calculate_metrics(data, ..., threshold = 2)
}
\arguments{
\item{data}{A dataframe containing the raw RT-QuIC data.}

\item{...}{A list of grouping factors.}

\item{threshold}{Float; the threshold applied to the calculation of time-to-threshold.}
}
\value{
A dataframe of calculated metrics.

#' @examples
file <- system.file(
"extdata/input_files",
file = "test.xlsx",
package = "quicR"
)
get_quic(file) |>
calculate_metrics(\verb{Sample IDs}, Dilutions, Wells)
}
\description{
Uses functions from the "calculate" family of quicR functions to generate an analyzed dataframe.
}
