% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MPR.R
\name{calculate_MPR}
\alias{calculate_MPR}
\title{Calculate the Maxpoint Ratio}
\usage{
calculate_MPR(data, col = "Norm", .by = "Well")
}
\arguments{
\item{data}{A data frame output from 'get_quic()'.}

\item{col}{The column containing the normalized fluorescence data.}

\item{.by}{Grouping factor. Should typically be by individual wells.}
}
\value{
A data frame containing well-matched MPR values.
}
\description{
Maxpoint ratio (MPR) is defined as the maximum relative fluorescence divided by the
background fluorescence.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_quic(file) |>
  calculate_MPR()

}
