% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MS.R
\name{calculate_MS}
\alias{calculate_MS}
\title{Calculate Maximum Slope}
\usage{
calculate_MS(data, start_col = 3)
}
\arguments{
\item{data}{A dataframe containing real-time reads. It is recommended to use a dataframe made from normalize_RFU.}

\item{start_col}{The column containing the starting position of the real-time data.}
}
\value{
A dataframe containing the real-time slope values.
}
\description{
Uses a sliding window to calculate the slope of real-time reads.
}
