% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MS.R
\name{calculate_MS}
\alias{calculate_MS}
\title{Calculate Maximum Slope}
\usage{
calculate_MS(data, window = 3)
}
\arguments{
\item{data}{A dataframe containing real-time reads. It is recommended to use a dataframe made from normalize_RFU.}

\item{window}{Integer designating how wide you want the sliding window to be for calculating the moving average slope.}
}
\value{
A dataframe containing the real-time slope values.
}
\description{
Uses a sliding window to calculate the slope of real-time reads.
}
