% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_QR.R
\name{calculate_QR}
\alias{calculate_QR}
\title{Calculate the Quenching Ratio}
\usage{
calculate_QR(
  data,
  col = "Norm",
  time_col = "Time",
  .by = "Wells",
  flip_ratio = FALSE
)
}
\arguments{
\item{data}{A data frame output from 'get_quic()'.}

\item{col}{The column containing the normalized fluorescence data.}

\item{time_col}{The column containing the time points.}

\item{.by}{Grouping factor. Should typically be by individual wells.}

\item{flip_ratio}{Logical; Should the ratio be calculated as max / last (default), or last / max?}
}
\value{
A data frame containing well-matched quenching ratio values.
}
\description{
The quenching ratio (QR) is defined as the ratio between the maximum initial value
after the exponential phase divided by the last measured value.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
get_quic(file) |>
  calculate_QR()

}
