% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TtT.R
\name{calculate_TtT}
\alias{calculate_TtT}
\title{Calculate Time to Threshold}
\usage{
calculate_TtT(
  data,
  threshold,
  time = "Time",
  values = "Norm",
  .by = lifecycle::deprecated(),
  by = "Well"
)
}
\arguments{
\item{data}{A dataframe containing real-time RT-QuIC data.}

\item{threshold}{A numeric value defining the threshold.}

\item{time}{Column containing your time values.}

\item{values}{Column containing your fluorescence values.}

\item{.by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use "by" instead.}

\item{by}{Grouping factor(s). Should typically be by individual wells. Can be supplied a vector as an argument.}
}
\value{
A vector containing the times to threshold
}
\description{
Calculates the time required to reach a defined threshold.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
get_quic(file) |>
  calculate_TtT(threshold = 3)

}
