% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TtT.R
\name{calculate_TtT}
\alias{calculate_TtT}
\title{Calculate Time to Threshold}
\usage{
calculate_TtT(data, threshold, time = "Time", values = "Norm", .by = "Wells")
}
\arguments{
\item{data}{A dataframe containing real-time RT-QuIC data.}

\item{threshold}{A numeric value defining the threshold.}

\item{time}{Column containing your time values.}

\item{values}{Column containing your fluorescence values.}

\item{.by}{Grouping factor. Should typically be by individual wells.}
}
\value{
A vector containing the times to threshold
}
\description{
Calculates the time required to reach a defined threshold.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test2.xlsx",
  package = "quicR"
)
get_quic(file) |>
  calculate_TtT(threshold = 2)

}
