% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tables.R
\name{convert_tables}
\alias{convert_tables}
\title{Convert tables into a single column in a dataframe.}
\usage{
convert_tables(tab)
}
\description{
convert_tables accepts a table or matrix or a list of tables and matrices
and converts them into dataframe columns.
}
\details{
@param tab A table/matrix or a list of tables/matrices.

@importFrom tidyr gather
@importFrom dplyr select

@export
}
