% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quic.R
\name{get_quic}
\alias{get_quic}
\title{Get all time-series and metadata.}
\usage{
get_quic(
  file,
  transpose_table = TRUE,
  normalize = TRUE,
  norm_point = 2,
  which_table = 1,
  .by = "Wells"
)
}
\arguments{
\item{file}{An Excel file exported by BMG.}

\item{transpose_table}{Logical, if true, will transpose the table(s). This should almost always be true.}

\item{normalize}{Logical, if true, will normalize the fluorescent data against the values at a certain cycle defined by 'norm_point'.}

\item{norm_point}{Integer, defines the cycle to use as background fluorescence.}

\item{which_table}{Integer, defines which table in the Excel sheet contains the real-time data. Should usually be set to 1.}

\item{.by}{Grouping factor. Should typically be by individual wells.}
}
\value{
A data frame containing all time-series data and sample/plate metadata.
}
\description{
Accepts an Excel file or a data frame of real-time RT-QuIC data.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_quic(file)

}
