% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real.R
\name{get_real}
\alias{get_real}
\title{Get Real-Time RT-QuIC Fluorescence Data}
\usage{
get_real(data, order_by_sample = FALSE, transpose_table = TRUE)
}
\arguments{
\item{data}{Either an Excel file or a dataframe.}

\item{order_by_sample}{Logical, if true, will organize the columns by sample ID rather than by well.}

\item{transpose_table}{Logical, if true, will transpose the table(s).}
}
\value{
A list of dataframes containing the formatted real-time data.
}
\description{
Accepts an Excel file or a dataframe of real-time RT-QuIC data.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_real(file)

}
