% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real.R
\name{get_real}
\alias{get_real}
\title{Get Real-Time RT-QuIC Fluorescence Data}
\usage{
get_real(file, ordered = FALSE)
}
\arguments{
\item{file}{Either an Excel file or a dataframe.}

\item{ordered}{Logical, if true, will organize the columns by sample ID.}
}
\value{
A list of dataframes containing the formatted real-time data.
}
\description{
Accepts an Excel file or a dataframe of real-time RT-QuIC data.
}
\examples{
file <- system.file(
"extdata/input_files",
file = "test.xlsx",
package = "quicR"
)
get_real(file)


}
