% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wells.R
\name{get_wells}
\alias{get_wells}
\title{Get the Wells Used in the RT-QuIC Run.}
\usage{
get_wells(file, sheet = 2)
}
\arguments{
\item{file}{Excel file exported from MARS}

\item{sheet}{The sheet containing the well IDs in the Excel file.}
}
\value{
A vector containing well IDs.
}
\description{
Returns the well IDs used in the plate.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test.xlsx",
  package = "quicR"
)
get_wells(file)

}
