% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_RFU.R
\name{normalize_RFU}
\alias{normalize_RFU}
\title{Normalize Fluorescence}
\usage{
normalize_RFU(data, bg_cycle = 4, transposed = FALSE)
}
\arguments{
\item{data}{A dataframe generated from get_real.}

\item{bg_cycle}{The cycle used for background fluorescence}

\item{transposed}{Logical, TRUE if cycle values are shown as column names.}
}
\value{
A dataframe containing real-time normalized fluorescence values.
}
\description{
Normalizes the real-time RT-QuIC data against the background fluorescence of
a defined cycle. All cycles are divided by the fluorescent value of the
defined cycle.
}
