% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_view.R
\name{plate_view}
\alias{plate_view}
\title{Real-Time Plate View}
\usage{
plate_view(
  data,
  plate = 96,
  sep = "\\n",
  plot_deriv = TRUE,
  flu_color = "black",
  der_color = "blue"
)
}
\arguments{
\item{data}{Real-time dataframe}

\item{plate}{Integer either 96 or 384 to denote microplate type.}

\item{sep}{A string defining how sample IDs and dilutions should be separated.}

\item{plot_deriv}{Logical; should the derivative be plotted?}

\item{flu_color}{Line color of the fluorescent data.}

\item{der_color}{Line color of the derivative data.}
}
\value{
A ggplot object
}
\description{
Converts the real-time data into a ggplot figure. The layout is either 8x12
or 16x24 for 96- and 384-well plates, respectively.
}
\examples{
\donttest{
file <- system.file(
  "extdata/input_files",
  file = "raw.csv",
  package = "quicR"
)

# Get the real-time data.
df_ <- read.csv(file, check.names=FALSE)

plate_view(df_)
}

}
