% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot metrics generated from the "calculate" family of quicR functions.}
\usage{
plot_metrics(data, ..., dilution_bool = TRUE, nrow = 2, ncol = 2)
}
\arguments{
\item{data}{A dataframe containing the calculated metrics from the "calculate" family of quicR functions.}

\item{...}{A selection of columns to be used as faceting variables. Typically this includes the sample IDs and dilution factors.}

\item{dilution_bool}{Logical; should dilution factors be included in the plot?}

\item{nrow}{Integer; number of rows to output in the plot.}

\item{ncol}{Integer; number of columns to output in the plot.}
}
\value{
A ggplot object
}
\description{
Generates a faceted figure of boxplots.
}
\examples{
file <- system.file(
  "extdata/input_files",
  file = "test4.xlsx",
  package = "quicR"
)

raw <- quicR::get_real(file)[[1]] |>
  quicR::normalize_RFU()

meta <- quicR::organize_tables(file) |>
  quicR::convert_tables()

data.frame(
  `Sample IDs` = meta$`Sample IDs`,
  `Dilutions`  = meta$Dilutions,
  check.names = FALSE
) |>
  dplyr::mutate(
    MPR = quicR::calculate_MPR(raw),
    MS  = quicR::calculate_MS(raw),
    TtT = quicR::calculate_TtT(raw, 2),
    RAF = 1 / TtT
  ) |>
  plot_metrics("Sample IDs", "Dilutions")

}
